% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readOxcal.R
\name{readOxcal}
\alias{readOxcal}
\title{Read OxCAL output files}
\usage{
readOxcal(x)
}
\arguments{
\item{x}{filename of the OxCAL \code{.js} file.}
}
\value{
a list
}
\description{
TODO
}
\details{
TODO
}
\author{
Gavin L. Simpson
}

